% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module1_cleaning_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean earthquake data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{Earthquake data.}
}
\value{
This function returns cleaned data with new date variable
and converts key variables used in package functions as numeric.
}
\description{
This function cleans earthquake data provided by
National Centers for Environmental Information.
}
\note{
Variable names as default YEAR, MONTH, DAY, LATITUDE, LONGITUDE, EQ_PRIMARY and DEATHS
}
\examples{
\dontrun{
readr::read_delim("signif.txt", delim = "\t") \%>\%
   eq_clean_data()
}

}
\references{
National Centers for Environmental Information,
National Oceanic and Atmospheric Administration (NOAA)
}
