% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module3_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Make label to leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{dataset with variables LOCATION_NAME, EQ_PRIMARY (magnitude)
and TOTAL_DEATHS for number of deaths}
}
\description{
This function creates leaflet map labels
}
\examples{
\dontrun{
# With several countries and popup text
data <- data.frame(LOCATION_NAME = "FINLAND", EQ_PRIMARY = "0.0", TOTAL_DEATHS = NA)
eq_create_label(data)
}
}
