% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module1_cleaning_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location name}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{Data with location name variable "LOCATION_NAME".}
}
\value{
This function returns location name as title case and without any text before ":".
}
\description{
This function cleans a location name.
}
\examples{
\dontrun{
data <- data.frame(LOCATION_NAME = "FINLAND:  KORVATUNTURI")
 eq_location_clean(data)
}

}
