% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module3_functions.R
\name{eq_map}
\alias{eq_map}
\title{Plot map with earthquakes}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Dataset with variables LOCATION_NAME, EQ_PRIMARY (magnitude)
and TOTAL_DEATHS for number of deaths.}

\item{annot_col}{Column name to be used for annotations.}
}
\value{
Map plot with interactive labels.
}
\description{
This function plots map with earthquakes. One or several countries
can be chosen.
}
\examples{
\dontrun{
# With several countries and popup text
readr::read_delim("signif.txt", delim = "\t") \%>\%
 eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "ITALY" | COUNTRY == "TURKEY" & lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
}.

}
