% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module2_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Make time line points}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{List of aesthetic created by aes() to use for plot.}

\item{data}{The data to be displayed in this layer. NULL, the default, is the data inherited from the plot data as specified in the call to ggplot().}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped.}

\item{inherit.aes}{Will aesthetics be inherited (TRUE/FALSE). If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to layer.}
}
\value{
PointsGrob object.
}
\description{
This function makes plot time line of earthquakes in selected time range with a point for each earthquake. Optional aesthetics include color, size, and alpha (for transparency). The xaesthetic is a date and an optional y aesthetic is a factor indicating some stratification in which case multiple time lines will be plotted for each level of the factor (e.g. country).
}
\examples{
\dontrun{
readr::read_delim("signif.txt", delim = "\t") \%>\%
eq_clean_data() \%>\%
 dplyr::filter((COUNTRY == "USA" ) & (lubridate::year(DATE) >= 2000 &
                                        lubridate::year(DATE) <= 2020)) \%>\%
 ggplot() +
 theme_eq + ggplot2::labs(size = "Richter scale value", fill = "# deaths") +
 geom_timeline( aes(x = DATE, size = EQ_PRIMARY, fill = DEATHS))
}

}
