% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module2_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Make time line labels}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  nudge_x = 0,
  nudge_y = 0,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max
)
}
\arguments{
\item{mapping}{List of aesthetic created by aes() to use for plot.}

\item{data}{The data to be displayed in this layer. NULL, the default, is the data inherited from the plot data as specified in the call to ggplot().}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment.}

\item{...}{Other arguments passed on to layer.}

\item{nudge_x}{Horizontal adjustment to nudge labels by.}

\item{nudge_y}{Vertical adjustment to nudge labels by.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped.}

\item{inherit.aes}{Will aesthethics be inherited (TRUE/FALSE). If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{n_max}{Number of labels, ordered by magnitude. Default is 5.}
}
\description{
This function adds annotations to the timeline plot. There is an option to subset to \code{n_max} to select the number of annotated earthquakes ordered by magnitude.
}
\examples{
\dontrun{
readr::read_delim("signif.txt", delim = "\t") \%>\%
 eq_clean_data() \%>\%
 dplyr::filter((COUNTRY == "USA" | COUNTRY == "CHINA") & (lubridate::year(DATE) >= 2000 &
                                                            lubridate::year(DATE) <= 2015)) \%>\%
 ggplot() +
 geom_timeline( aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, fill = DEATHS)) +
 geom_timeline_label( aes(x = DATE, y = COUNTRY, label = LOCATION_NAME,
 n_max = 1, size = EQ_PRIMARY)) +
 theme_eq
}

}
