% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module2_functions.R
\docType{data}
\name{theme_eq}
\alias{theme_eq}
\title{Modify plot theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_eq
}
\description{
Plot theme is modified from theme_minimal from qqplot2 package.
}
\keyword{datasets}
